@@ -1,10 +1,8 @@
 ﻿using DnDGen.Infrastructure.Mappers.Collections;
-using DnDGen.Infrastructure.Tables;
 using Moq;
 using NUnit.Framework;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 
 namespace DnDGen.Infrastructure.Tests.Unit.Mappers.Collections
@@ -15,38 +13,54 @@ namespace DnDGen.Infrastructure.Tests.Unit.Mappers.Collections
         private CollectionMapper proxy;
         private Mock<CollectionMapper> mockInnerMapper;
         private Dictionary<string, IEnumerable<string>> table;
-        private Mock<AssemblyLoader> mockAssemblyLoader;
 
         [SetUp]
         public void Setup()
         {
             mockInnerMapper = new Mock<CollectionMapper>();
-            mockAssemblyLoader = new Mock<AssemblyLoader>();
-            proxy = new CollectionMapperCachingProxy(mockInnerMapper.Object, mockAssemblyLoader.Object);
+            proxy = new CollectionMapperCachingProxy(mockInnerMapper.Object);
 
             table = new Dictionary<string, IEnumerable<string>>();
             table["name"] = new[] { "entry 1", "entry 2" };
             table["other name"] = new[] { "entry 3", "entry 4" };
 
-            mockInnerMapper.Setup(m => m.Map("table name")).Returns(table);
-            mockAssemblyLoader.Setup(l => l.GetRunningAssembly()).Returns(Assembly.GetExecutingAssembly());
+            mockInnerMapper.Setup(m => m.Map("my assembly", "table name")).Returns(table);
         }
 
         [Test]
         public void ReturnTableFromInnerMapper()
         {
-            var result = proxy.Map("table name");
+            var result = proxy.Map("my assembly", "table name");
             Assert.That(result, Is.EqualTo(table));
         }
 
         [Test]
         public void CacheTable()
         {
-            proxy.Map("table name");
-            var result = proxy.Map("table name");
+            proxy.Map("my assembly", "table name");
+            var result = proxy.Map("my assembly", "table name");
 
             Assert.That(result, Is.EqualTo(table));
-            mockInnerMapper.Verify(p => p.Map(It.IsAny<string>()), Times.Once);
+            mockInnerMapper.Verify(p => p.Map(It.IsAny<string>(), It.IsAny<string>()), Times.Once);
+        }
+
+        [Test]
+        public void DoNotCacheDifferentTable()
+        {
+            var otherTable = new Dictionary<string, IEnumerable<string>>();
+            otherTable["other name"] = new[] { "other entry 1", "other entry 2" };
+            otherTable["other other name"] = new[] { "other entry 3", "other entry 4" };
+
+            mockInnerMapper.SetupSequence(m => m.Map("my assembly", "other table name")).Returns(otherTable);
+
+            var result = proxy.Map("my assembly", "table name");
+            var otherResult = proxy.Map("my assembly", "other table name");
+
+            Assert.That(result, Is.EqualTo(table));
+            Assert.That(otherResult, Is.EqualTo(otherTable));
+            mockInnerMapper.Verify(p => p.Map(It.IsAny<string>(), It.IsAny<string>()), Times.Exactly(2));
+            mockInnerMapper.Verify(p => p.Map("my assembly", "table name"), Times.Once);
+            mockInnerMapper.Verify(p => p.Map("my assembly", "other table name"), Times.Once);
         }
 
         [Test]
@@ -55,7 +69,7 @@ namespace DnDGen.Infrastructure.Tests.Unit.Mappers.Collections
             var tasks = new List<Task<Dictionary<string, IEnumerable<string>>>>();
             while (tasks.Count < 10)
             {
-                var task = Task.Run(() => proxy.Map("table name"));
+                var task = Task.Run(() => proxy.Map("my assembly", "table name"));
                 tasks.Add(task);
             }
 
@@ -67,7 +81,7 @@ namespace DnDGen.Infrastructure.Tests.Unit.Mappers.Collections
                 Assert.That(result, Is.EqualTo(table));
             }
 
-            mockInnerMapper.Verify(p => p.Map(It.IsAny<string>()), Times.Once);
+            mockInnerMapper.Verify(p => p.Map(It.IsAny<string>(), It.IsAny<string>()), Times.Once);
         }
 
         [Test]
@@ -77,28 +91,21 @@ namespace DnDGen.Infrastructure.Tests.Unit.Mappers.Collections
             otherTable["other name"] = new[] { "other entry 1", "other entry 2" };
             otherTable["other other name"] = new[] { "other entry 3", "other entry 4" };
 
-            mockInnerMapper.SetupSequence(m => m.Map("table name"))
-                .Returns(table)
-                .Returns(otherTable)
-                .Returns(table)
-                .Returns(otherTable);
+            mockInnerMapper.SetupSequence(m => m.Map("my assembly", "table name")).Returns(table);
+            mockInnerMapper.SetupSequence(m => m.Map("my other assembly", "table name")).Returns(otherTable);
 
-            mockAssemblyLoader.SetupSequence(l => l.GetRunningAssembly())
-                .Returns(Assembly.GetExecutingAssembly())
-                .Returns(typeof(int).Assembly)
-                .Returns(Assembly.GetExecutingAssembly())
-                .Returns(typeof(int).Assembly);
+            proxy.Map("my assembly", "table name");
+            proxy.Map("my other assembly", "table name");
 
-            proxy.Map("table name");
-            proxy.Map("table name");
-
-            var firstResult = proxy.Map("table name");
-            var secondResult = proxy.Map("table name");
+            var firstResult = proxy.Map("my assembly", "table name");
+            var secondResult = proxy.Map("my other assembly", "table name");
 
             Assert.That(firstResult, Is.EqualTo(table));
             Assert.That(secondResult, Is.EqualTo(otherTable));
 
-            mockInnerMapper.Verify(p => p.Map(It.IsAny<string>()), Times.Exactly(2));
+            mockInnerMapper.Verify(p => p.Map(It.IsAny<string>(), It.IsAny<string>()), Times.Exactly(2));
+            mockInnerMapper.Verify(p => p.Map("my assembly", It.IsAny<string>()), Times.Once);
+            mockInnerMapper.Verify(p => p.Map("my other assembly", It.IsAny<string>()), Times.Once);
         }
     }
 }
