@@ -1,5 +1,4 @@
-﻿using DnDGen.Infrastructure.Tables;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace DnDGen.Infrastructure.Mappers.Collections
 {
@@ -7,28 +6,25 @@ namespace DnDGen.Infrastructure.Mappers.Collections
     {
         private readonly CollectionMapper innerMapper;
         private readonly Dictionary<string, Dictionary<string, IEnumerable<string>>> cachedTables;
-        private readonly AssemblyLoader assemblyLoader;
         private readonly object myLock;
 
-        public CollectionMapperCachingProxy(CollectionMapper innerMapper, AssemblyLoader assemblyLoader)
+        public CollectionMapperCachingProxy(CollectionMapper innerMapper)
         {
             this.innerMapper = innerMapper;
-            this.assemblyLoader = assemblyLoader;
 
             cachedTables = new Dictionary<string, Dictionary<string, IEnumerable<string>>>();
             myLock = new object();
         }
 
-        public Dictionary<string, IEnumerable<string>> Map(string tableName)
+        public Dictionary<string, IEnumerable<string>> Map(string assemblyName, string tableName)
         {
-            var assembly = assemblyLoader.GetRunningAssembly();
-            var key = assembly.FullName + tableName;
+            var key = assemblyName + tableName;
 
             lock (myLock)
             {
                 if (!cachedTables.ContainsKey(key))
                 {
-                    var mappedTable = innerMapper.Map(tableName);
+                    var mappedTable = innerMapper.Map(assemblyName, tableName);
                     cachedTables.Add(key, mappedTable);
                 }
             }
