@@ -4,6 +4,6 @@ namespace DnDGen.Infrastructure.Mappers.Percentiles
 {
     public interface PercentileMapper
     {
-        Dictionary<int, string> Map(string tableName);
+        Dictionary<int, string> Map(string assemblyName, string tableName);
     }
 }
