@@ -1,20 +1,17 @@
-﻿using DnDGen.Infrastructure.Tables;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace DnDGen.Infrastructure.Selectors.Collections
 {
     internal class CollectionSelectorCachingProxy : ICollectionSelector
     {
         private readonly ICollectionSelector innerSelector;
-        private readonly AssemblyLoader assemblyLoader;
         private readonly Dictionary<string, IEnumerable<string>> cachedExplodedCollections;
         private readonly Dictionary<string, IEnumerable<string>> cachedExplodedWithDuplicatesCollections;
         private readonly object myLock;
 
-        public CollectionSelectorCachingProxy(ICollectionSelector innerSelector, AssemblyLoader assemblyLoader)
+        public CollectionSelectorCachingProxy(ICollectionSelector innerSelector)
         {
             this.innerSelector = innerSelector;
-            this.assemblyLoader = assemblyLoader;
 
             cachedExplodedCollections = new Dictionary<string, IEnumerable<string>>();
             cachedExplodedWithDuplicatesCollections = new Dictionary<string, IEnumerable<string>>();
@@ -26,16 +23,15 @@ namespace DnDGen.Infrastructure.Selectors.Collections
             return innerSelector.CreateWeighted(common, uncommon, rare, veryRare);
         }
 
-        public IEnumerable<string> Explode(string tableName, string collectionName)
+        public IEnumerable<string> Explode(string assemblyName, string tableName, string collectionName)
         {
-            var assembly = assemblyLoader.GetRunningAssembly();
-            var key = assembly.FullName + tableName + collectionName;
+            var key = assemblyName + tableName + collectionName;
 
             lock (myLock)
             {
                 if (!cachedExplodedCollections.ContainsKey(key))
                 {
-                    var explodedCollection = innerSelector.Explode(tableName, collectionName);
+                    var explodedCollection = innerSelector.Explode(assemblyName, tableName, collectionName);
                     cachedExplodedCollections.Add(key, explodedCollection);
                 }
             }
@@ -43,16 +39,15 @@ namespace DnDGen.Infrastructure.Selectors.Collections
             return cachedExplodedCollections[key];
         }
 
-        public IEnumerable<string> ExplodeAndPreserveDuplicates(string tableName, string collectionName)
+        public IEnumerable<string> ExplodeAndPreserveDuplicates(string assemblyName, string tableName, string collectionName)
         {
-            var assembly = assemblyLoader.GetRunningAssembly();
-            var key = assembly.FullName + tableName + collectionName;
+            var key = assemblyName + tableName + collectionName;
 
             lock (myLock)
             {
                 if (!cachedExplodedWithDuplicatesCollections.ContainsKey(key))
                 {
-                    var explodedCollection = innerSelector.ExplodeAndPreserveDuplicates(tableName, collectionName);
+                    var explodedCollection = innerSelector.ExplodeAndPreserveDuplicates(assemblyName, tableName, collectionName);
                     cachedExplodedWithDuplicatesCollections.Add(key, explodedCollection);
                 }
             }
@@ -60,9 +55,9 @@ namespace DnDGen.Infrastructure.Selectors.Collections
             return cachedExplodedWithDuplicatesCollections[key];
         }
 
-        public string FindCollectionOf(string tableName, string entry, params string[] filteredCollectionNames)
+        public string FindCollectionOf(string assemblyName, string tableName, string entry, params string[] filteredCollectionNames)
         {
-            return innerSelector.FindCollectionOf(tableName, entry, filteredCollectionNames);
+            return innerSelector.FindCollectionOf(assemblyName, tableName, entry, filteredCollectionNames);
         }
 
         public IEnumerable<string> Flatten(Dictionary<string, IEnumerable<string>> collections, IEnumerable<string> keys)
@@ -70,19 +65,19 @@ namespace DnDGen.Infrastructure.Selectors.Collections
             return innerSelector.Flatten(collections, keys);
         }
 
-        public bool IsCollection(string tableName, string collectionName)
+        public bool IsCollection(string assemblyName, string tableName, string collectionName)
         {
-            return innerSelector.IsCollection(tableName, collectionName);
+            return innerSelector.IsCollection(assemblyName, tableName, collectionName);
         }
 
-        public Dictionary<string, IEnumerable<string>> SelectAllFrom(string tableName)
+        public Dictionary<string, IEnumerable<string>> SelectAllFrom(string assemblyName, string tableName)
         {
-            return innerSelector.SelectAllFrom(tableName);
+            return innerSelector.SelectAllFrom(assemblyName, tableName);
         }
 
-        public IEnumerable<string> SelectFrom(string tableName, string collectionName)
+        public IEnumerable<string> SelectFrom(string assemblyName, string tableName, string collectionName)
         {
-            return innerSelector.SelectFrom(tableName, collectionName);
+            return innerSelector.SelectFrom(assemblyName, tableName, collectionName);
         }
 
         public T SelectRandomFrom<T>(IEnumerable<T> collection)
@@ -90,9 +85,9 @@ namespace DnDGen.Infrastructure.Selectors.Collections
             return innerSelector.SelectRandomFrom(collection);
         }
 
-        public string SelectRandomFrom(string tableName, string collectionName)
+        public string SelectRandomFrom(string assemblyName, string tableName, string collectionName)
         {
-            return innerSelector.SelectRandomFrom(tableName, collectionName);
+            return innerSelector.SelectRandomFrom(assemblyName, tableName, collectionName);
         }
 
         public T SelectRandomFrom<T>(IEnumerable<T> common = null, IEnumerable<T> uncommon = null, IEnumerable<T> rare = null, IEnumerable<T> veryRare = null)
