@@ -1,23 +1,12 @@
-﻿using System.Diagnostics;
-using System.Linq;
-using System.Reflection;
+﻿using System.Reflection;
 
 namespace DnDGen.Infrastructure.Tables
 {
     internal class DomainAssemblyLoader : AssemblyLoader
     {
-        public Assembly GetRunningAssembly()
+        public Assembly GetAssembly(string name)
         {
-            var stacktrace = new StackTrace();
-            var frames = stacktrace.GetFrames();
-            var assembly = frames
-                .Select(f => f.GetMethod())
-                .Select(m => m.ReflectedType.Assembly)
-                .First(a => !a.FullName.StartsWith("DnDGen.Infrastructure, ")
-                    && !a.FullName.StartsWith("DnDGen.Infrastructure.Tests.Unit, ")
-                    && !a.FullName.StartsWith("System"));
-
-            return assembly;
+            return Assembly.Load(name);
         }
     }
 }
